novoDiario ok, metatags de segurança ok

+++ controle/main.js
@@ -131,11 +131,13 @@ ipcMain.handle('salvar-feitico', async (_, nomeArquivoAntigo, novoTitulo, novoCo
 ipcMain.handle('salvar-diario', async(_, nomeArquivoAntigo, novoTitulo, novoConteudo)=>{
     try {
         const novoNomeArquivo=`${novoTitulo}.txt`
-        const caminhoAntigo = path.join(pastaDiarios, nomeArquivoAntigo)
         const caminhoNovo = path.join(pastaDiarios, novoNomeArquivo)
-
-        if (nomeArquivoAntigo!==novoNomeArquivo){
-            fs.renameSync(caminhoAntigo, caminhoNovo)
+        
+        if (nomeArquivoAntigo){
+            const caminhoAntigo = path.join(pastaDiarios, nomeArquivoAntigo)
+            if (nomeArquivoAntigo !== novoNomeArquivo && fs.existsSync(caminhoAntigo)){
+                fs.renameSync(caminhoAntigo, caminhoNovo)
+            }
         }
 
         fs.writeFileSync(caminhoNovo, novoConteudo, 'utf-8')
